requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.0
pyautogui>=0.9.54
pyperclip>=1.8.2
pyyaml>=6.0
keyboard>=0.13.5
pyinstaller>=6.0